Test fb_fill_pixels using sieve of Eratosthenes (compute primes below 100k)
There are exactly 9.592 prime numbers below 100.000.

This script uses fb_fill_pixels and video ram, and will test that the
correct number of pixels are not written to.

As a bonus, this happens in video ram, and its operation is visible in real
time. This script doubles as an interactive demo/visualization of this
simple yet effective prime number algorithm.

Note that counting pixels could have been done using peek/vpeek, but that
turns out to take more than a minute. Thus, a machine code function is
added for convenience.