Proof of Concept for resolving the clash of UDTIM with the 65C816's NMIB vector.
At $FFEA in the ROM is the jump table entry for UDTIM aka `jmp clock_update`.
In native mode, the 65C816 reads its NMIB vector from $FFEA as well. By
hardcoding the address for `clock_update`, which is an alias for
`softclock_timer_update`, to an address with the low byte $XX, the NMIB
handler can be hardcoded to $XX4C, as $4C is the opcode for an absolute `jmp`.
This PoC uses $EEEE for `softclock_timer_update` and $EE4C for the NMIB handler
as example values.