[KERNAL] Remove leftover RS-232 support vestiges (#335)
The RS-232 driver has not been doing anything for a while now, and the `brk`
in BASIN could cause seemingly working programs to break into the monitor
for no apparent reason. As the official hardware does not include RS-232,
and the existing support is incomplete and non-functional, it is better to
remove it entirely and make it error out like the tape device does.