[DOS/FAT32] Use carry for instead of `krn_ptr1` to indicate writes to the same address
`fat32_read` and `fat32_write` currently use `krn_ptr1` to determine whether to increment
the buffer pointer or not; the latter is used for reads from / writes to an IO address.
As this is a DOS variable with a variable address that is unrelated to the other named
FAT32 variables, this commit changes this to use the carry flag instead; calling the function
with carry set prevents address incrementing.