[BUILD] Fail early if findsymbols fails
If `findsymbols` fails, it prints a helpful usage message to standard
output:

> Usage: findsymbols file [-p prefix] symbol ...
>   file:
>     Symbol file created with the -Ln option in ld65
>   -p prefix:
>     A prefix added to all returned symbol names
>   symbol:
>     Name of one or more symbols

Unfortunately, the Makefile ignores the exit status of `findsymbols`. It
then misinterprets the whole help text as legitimate output and rolls it
all into the linker command line.

Because the help text happens to contain the word `-Ln` and there’s
another `-Ln` in the command line already, so the linker treats this
duplication as an error and bails:

> ld65: Error: Cannot use -Ln twice

which was also reported in GitHub issue X16Community/x16-rom#363 [1].

To improve the error message and fix the leaky help text:

1. Add a conditional block that queries `.SHELLSTATUS`, a variable which
was introduced in GNU Make v4.2, and if it is nonzero, error out early
with the message "findsymbols failed with exit code $(.SHELLSTATUS)."

2. Additionally, improve `findsymbols` so it redirects the usage text to
standard error and includes the file name in its "File not found"
message.
That way, the message can be seen by the user who ran `make`. It can
also no longer be mistaken for machine-readable output.

Example output of `make`:

> File not found: build/x16/dos.sym
> Usage: findsymbols file [-p prefix] symbol ...
>   file:
>     Symbol file created with the -Ln option in ld65
>   -p prefix:
>     A prefix added to all returned symbol names
>   symbol:
>     Name of one or more symbols
> Makefile:407: *** findsymbols failed with exit code 1.  Stop.

3. To avoid peppering the whole Makefile with exit code guards, declare
a `findsymbols` function that hides the complexity.

4. As an unwanted side effect, calling a function causes Make to
misinterpret the `$` address prefixes that occur in the `findsymbols`
output. To counteract that, escape every `$` with `$$`.

5. As a graceful fallback for users of GNU Make versions older than 4.2,
default `.SHELLSTATUS` to 0. In that case, we have no early returns, but
users still get a somewhat meaningful error message that says e.g.,
"Unresolved external 'skip_mask' referenced in: fat32/match.s", which is
still much more helpful than "ld65: Error: Cannot use -Ln twice".

[1]: https://github.com/X16Community/x16-rom/issues/363